+++ src/api/individualNotice/indivNoticeApi.js
@@ -71,7 +71,6 @@ export const postIndNotice = async ({
   }
 };
 
-
 // 댓글 등록하기
 export const postComment = async ({ product, successFn, failFn, errorFn }) => {
   try {
@@ -111,11 +110,14 @@ export const deleteComment = async ({
       failFn();
     }
   } catch (error) {
+    errorFn(error);
+  }
+};
 
 // 알림장 게시글 삭제하기
 export const deleteNotice = async ({ ikid, successFn, failFn, errorFn }) => {
   try {
-    const res = await jwtAxios.delete(`${path}?iteacher=1&ifullNotice=${ikid}`);
+    const res = await jwtAxios.delete(`${path}?inotice=${ikid}`);
     const status = res.status.toString();
 
     if (status.charAt(0) === "2") {
@@ -125,6 +127,5 @@ export const deleteNotice = async ({ ikid, successFn, failFn, errorFn }) => {
     }
   } catch (error) {
     errorFn(error);
-
   }
 };
